fix: validate goavro codec building (#43)
There is one case (that we've found) where a avro schema can be
technically valid but we are unable to build a codec using goavro.

Specifically, we can specify a default value that does not match the
type declaration and verifySchema will not catch the issue.  It will be
a runtime error when we try to build the codec later.

The failure case is something like this:
```
{
  "name": "Wizard",
  "type": "record",
  "namespace": "com.avro.kedavro",
  "fields": [{
    "name": "name",
    "type": "string",
    "default": null
  }]
 }
```

The in the case above the format of the avro is correct but when it
comes time to coerce the default value (null) into the field the the
type "string" is not nullable.

This commit causes this to throw an error.